Time: 41 ms (86.41%), Space: 6 MB (47.83%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int concatenatedBinary(int n) {
         int MOD=1e9+7;
-        long long res=0;
+        long res=0;
         for(int i=1;i<=n;i++){
             int len  = 32 - __builtin_clz(i);
             res = ((res<<len)|i)%MOD;
